Secured all api routes

+++ src/pages/api/deleteAccount.ts
@@ -1,10 +1,12 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import prisma from 'lib/prismadb';
+import { getSession } from 'next-auth/react';
 
 export default async function deleteAccount(req: NextApiRequest, res: NextApiResponse) {
   const { userId } = req.body;
+  const session = await getSession({ req });
 
-  if (userId) {
+  if (session && userId) {
     await prisma.user.delete({
       where: {
         id: userId,
