+++ src/pages/api/deleteMealsSection.ts
@@ -1,10 +1,12 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import prisma from 'lib/prismadb';
+import { getSession } from 'next-auth/react';
 
 export default async function deleteMealsSection(req: NextApiRequest, res: NextApiResponse) {
   const { dayId, mealsSectionId } = req.body;
+  const session = await getSession({ req });
 
-  if (dayId && mealsSectionId) {
+  if (session && dayId && mealsSectionId) {
     await prisma.mealsSection.delete({
       where: {
         id: mealsSectionId,
